+++ src/modules-arch/pseudo-modules/deposit-form-pseudo-module.js
@@ -1,5 +1,5 @@
 import { ModuleDescriptor } from '@/modules-arch/module-descriptor'
-import { CoinpaymentsDepositModule } from '@/vue/modules/coinpayments/module'
+import { CoinpaymentsDepositModule } from '@/vue/modules/coinpayments-deposit/module'
 
 export class DepositFormPseudoModule extends ModuleDescriptor {
   constructor (opts = {}) {